fix(contigo): un solo dueño del mapa de hábitos — las pantallas dejaban de verse entre sí
useContigoHabits guardaba el mapa completo desde el estado propio de cada
instancia; con las 4 pantallas de Contigo montadas a la vez, escribir en
una sin remontar la otra pisaba el mapa entero con una copia vieja y
des-completaba en silencio un hábito ya marcado.

ContigoHabitsProvider pasa a ser el único dueño del estado (montado en
contigo/_layout.tsx). Las mutaciones parten de recordsRef.current (nunca
de un closure de render viejo) y la persistencia se serializa con una cola
de promesas. useContigoHabits queda como fachada de una línea; las 4
pantallas no cambian.

Plan 004 de plans/README.md — marcado DONE.
