+++ mcm-app/app/(tabs)/contigo/_layout.tsx
@@ -1,44 +1,47 @@
 import React from 'react';
 import { Stack } from 'expo-router';
+import { ContigoHabitsProvider } from '@/contexts/ContigoHabitsContext';
 
 export default function ContigoLayout() {
   return (
-    <Stack initialRouteName="index">
-      <Stack.Screen
-        name="index"
-        options={{
-          headerShown: false,
-          title: 'Contigo',
-        }}
-      />
-      <Stack.Screen
-        name="bookmarks"
-        options={{
-          headerShown: false,
-          title: 'Favoritos',
-        }}
-      />
-      <Stack.Screen
-        name="evangelio"
-        options={{
-          headerShown: false,
-          title: 'Evangelio',
-        }}
-      />
-      <Stack.Screen
-        name="oracion"
-        options={{
-          headerShown: false,
-          title: 'Oración',
-        }}
-      />
-      <Stack.Screen
-        name="revision"
-        options={{
-          headerShown: false,
-          title: 'Revisión del día',
-        }}
-      />
-    </Stack>
+    <ContigoHabitsProvider>
+      <Stack initialRouteName="index">
+        <Stack.Screen
+          name="index"
+          options={{
+            headerShown: false,
+            title: 'Contigo',
+          }}
+        />
+        <Stack.Screen
+          name="bookmarks"
+          options={{
+            headerShown: false,
+            title: 'Favoritos',
+          }}
+        />
+        <Stack.Screen
+          name="evangelio"
+          options={{
+            headerShown: false,
+            title: 'Evangelio',
+          }}
+        />
+        <Stack.Screen
+          name="oracion"
+          options={{
+            headerShown: false,
+            title: 'Oración',
+          }}
+        />
+        <Stack.Screen
+          name="revision"
+          options={{
+            headerShown: false,
+            title: 'Revisión del día',
+          }}
+        />
+      </Stack>
+    </ContigoHabitsProvider>
   );
 }
